+++ fluent_contents/conf/plugin_template/content_plugins.py
@@ -1,0 +1,12 @@
+from fluent_contents.extensions import plugin_pool, ContentPlugin
+from .models import {{ model }}
+
+
+@plugin_pool.register
+class {{ plugin }}(ContentPlugin):
+    """
+    CMS plugin for ...
+    """
+    model = {{ model }}
+    #category = ContentPlugin.ADVANCED
+    render_template = "includes/fluentcms_plugins/{{ app_name }}.html"